refactor(format): move analysis renderer tests to owner module
